fixes types

+++ src/MultiSigWallet.sol
@@ -2,51 +2,51 @@
 pragma solidity ^0.8.13;
 
 contract MultiSigWallet {
-    event Deposit(address indexed sender, uint amount, uint balance);
-    event Initiated(uint indexed txId, address indexed owner, address indexed to, uint value, bytes data);
-    event Approved(uint indexed txId, address indexed owner);
-    event Executed(uint indexed txId, address indexed owner);
-    event Revoked(uint indexed txId, address indexed owner);
+    event Deposit(address indexed sender, uint256 amount, uint256 balance);
+    event Initiated(uint256 indexed txId, address indexed owner, address indexed to, uint256 value, bytes data);
+    event Approved(uint256 indexed txId, address indexed owner);
+    event Executed(uint256 indexed txId, address indexed owner);
+    event Revoked(uint256 indexed txId, address indexed owner);
 
     struct Transaction {
         address to;
-        uint value;
+        uint256 value;
         bytes data;
         bool executed;
     }
 
     address[] public owners;
-    uint public approvalsRequired;
+    uint256 public approvalsRequired;
     Transaction[] public transactions;
 
     mapping(address => bool) public isOwner;
-    mapping(uint => mapping(address => bool)) public approvals;
+    mapping(uint256 => mapping(address => bool)) public approvals;
 
     modifier onlyOwner() {
         require(isOwner[msg.sender], "Not owner");
         _;
     }
 
-    modifier exists(uint _txId) {
+    modifier exists(uint256 _txId) {
         require(_txId < transactions.length, "Transaction does not exist");
         _;
     }
 
-    modifier notApproved(uint _txId) {
+    modifier notApproved(uint256 _txId) {
         require(!approvals[_txId][msg.sender], "Already approved");
         _;
     }
 
-    modifier notExecuted(uint _txId) {
+    modifier notExecuted(uint256 _txId) {
         require(!transactions[_txId].executed, "Already executed");
         _;
     }
 
-    constructor(address[] memory _owners, uint _approvalsRequired) {
+    constructor(address[] memory _owners, uint256 _approvalsRequired) {
         require(_owners.length > 0, "Not enough owners");
         require(_approvalsRequired > 0 && _approvalsRequired <= _owners.length, "Invalid amount of approvals");
 
-        for (uint i; i < _owners.length; i++) {
+        for (uint256 i; i < _owners.length; i++) {
             address owner = _owners[i];
             
             require(owner != address(0), "Invalid owner");
@@ -59,17 +59,17 @@ contract MultiSigWallet {
         approvalsRequired = _approvalsRequired;
     }
 
-    function initiate(address _to, uint _value, bytes calldata _data) public onlyOwner {
+    function initiate(address _to, uint256 _value, bytes calldata _data) public onlyOwner {
         transactions.push(Transaction(_to, _value, _data, false));
         emit Initiated(transactions.length - 1, msg.sender, _to, _value, _data);
     }
 
-    function approve(uint _txId) public onlyOwner exists(_txId) notApproved(_txId) notExecuted(_txId) {
+    function approve(uint256 _txId) public onlyOwner exists(_txId) notApproved(_txId) notExecuted(_txId) {
         approvals[_txId][msg.sender] = true;
         emit Approved(_txId, msg.sender);
     }
 
-    function execute(uint _txId) public onlyOwner exists(_txId) notExecuted(_txId) {
+    function execute(uint256 _txId) public onlyOwner exists(_txId) notExecuted(_txId) {
         require(_getApprovalCount(_txId) >= approvalsRequired, "Insufficient approvals");
         require(address(this).balance >= transactions[_txId].value, "Insufficient balance");
         
@@ -82,7 +82,7 @@ contract MultiSigWallet {
         emit Executed(_txId, msg.sender);
     }
 
-    function revoke(uint _txId) public onlyOwner exists(_txId) notExecuted(_txId) {
+    function revoke(uint256 _txId) public onlyOwner exists(_txId) notExecuted(_txId) {
         approvals[_txId][msg.sender] = false;
         emit Revoked(_txId, msg.sender);
     }
@@ -95,20 +95,20 @@ contract MultiSigWallet {
         return owners;
     }
 
-    function getApprovalsRequired() public view returns (uint) {
+    function getApprovalsRequired() public view returns (uint256) {
         return approvalsRequired;
     }
 
-    function getTransactionCount() public view returns (uint) {
+    function getTransactionCount() public view returns (uint256) {
         return transactions.length;
     }
 
-    function getApprovalStatus(uint _txId) public view returns (bool) {
+    function getApprovalStatus(uint256 _txId) public view returns (bool) {
         return approvals[_txId][msg.sender];
     }
 
-    function _getApprovalCount(uint _txId) private view returns (uint count) {
-        for (uint i; i < owners.length; i++) {
+    function _getApprovalCount(uint256 _txId) private view returns (uint256 count) {
+        for (uint256 i; i < owners.length; i++) {
             if (approvals[_txId][owners[i]]) {
                 count++;
             }
